requirements.txt
# Enhanced AI Forensics - Python Dependencies
# Install these packages for the enhanced PyTorch detector

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.8.0
pillow>=8.0.0
matplotlib>=3.5.0
albumentations>=1.3.0
scikit-learn>=1.0.0